Feat: 경매 중인 상품은 삭제 불가능 예외 처리 추가

+++ server/wonprice/src/main/java/main/wonprice/domain/product/controller/ProductController.java
@@ -137,15 +137,23 @@ public class ProductController {
     @DeleteMapping("/{productId}")
     public ResponseEntity deleteProduct(@PathVariable Long productId) {
         Member loginMember = memberService.findLoginMember();
-        Product product = productService.deleteOneById(productId, loginMember);
+        Product product = productService.findOneById(productId);
 
         Long productOwnerId = product.getSeller().getMemberId();
         Long loginMemberId = loginMember.getMemberId();
 
         // 상품 판매자와 로그인 한 사용자가 다를 경우, 권한이 없음을 응답
         if (!productOwnerId.equals(loginMemberId)) {
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("게시글을 삭제 할 권한이 없습니다.");
         }
+
+        // 경매 상품인 경우, buyer_id가 있는지 확인하여 삭제 여부 결정
+        if(product.getAuction() && product.getBuyerId() != null){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("경매가 진행 중인 상품은 삭제할 수 없습니다.");
+        }
+
+        // 삭제가 가능한 경우, 상품 삭제
+        productService.deleteOneById(productId, loginMember);
 
         return ResponseEntity.noContent().build();
     }
